listener, udp-listener: don't delete abstract sockets

+++ src/listener.c
@@ -118,7 +118,9 @@ listener_new(struct pool *pool, int family, int socktype, int protocol,
 
     if (address->sa_family == AF_UNIX) {
         const struct sockaddr_un *sun = (const struct sockaddr_un *)address;
-        unlink(sun->sun_path);
+        if (sun->sun_path[0] != '\0')
+            /* delete non-abstract socket files before reusing them */
+            unlink(sun->sun_path);
     }
 
     ret = bind(listener->fd, address, address_length);
